fix: Address critical issues from BUGS.md and improve code quality
Major fixes:
- Implement configuration management with JSON config file support
- Fix memoryEngine.Close() nil pointer dereference issue
- Address linting issues and improve code quality
- Clean up outdated documentation and organize planning docs

Configuration:
- Add internal/config package with comprehensive configuration system
- Support JSON config files with command-line flag overrides
- Create example config file (quantadb.json.example)
- Update main.go to load and use configuration

Storage & Engine:
- Fix memoryEngine.Close() to handle closed state properly
- Add ErrEngineClosed error constant
- Replace context.Background() with context.TODO() and documentation
- Add comprehensive close handling tests

Code Quality:
- Fix unreachable code in partial_index.go
- Add nolint directives for safe integer conversions
- Fix unused fields and parameters
- Improve error handling patterns
- Pre-allocate slices where beneficial
- Fix code style issues (indent-error-flow)

Documentation & Organization:
- Archive completed Phase 3 and Phase 4 planning documents
- Update CURRENT_STATUS.md with Phase 4 completion
- Create Phase 5 distributed planning document
- Add TODO.md files for placeholder commands
- Move remaining BUGS.md items to TODO.md
- Delete BUGS.md after consolidation

Testing:
- Add memory_close_test.go for engine close handling
- Ensure all tests pass after changes

+++ internal/storage/page.go
@@ -126,5 +126,5 @@ func (p *Page) HasSpaceFor(size uint16) bool {
 	dataStart := p.Header.FreeSpacePtr - size - PageHeaderSize
 
 	// Ensure slot array won't overlap with data area
-	return uint16(nextSlotEnd) < dataStart
+	return nextSlotEnd < dataStart
 }
